refactor account route structure
