+++ apps/app/src/routes/_authenticated/account/settings.tsx
@@ -1,5 +1,5 @@
-import { createFileRoute, Outlet, redirect } from "@tanstack/react-router";
-import { SettingsSidebar } from "~/components/settings-sidebar";
+import { createFileRoute, redirect } from "@tanstack/react-router";
+import { AccountSettingsLayout } from "~/account/settings/account-settings-layout";
 
 export const Route = createFileRoute("/_authenticated/account/settings")({
   beforeLoad: ({ location }) => {
@@ -21,33 +21,3 @@ export const Route = createFileRoute("/_authenticated/account/settings")({
   }),
   component: AccountSettingsLayout,
 });
-
-function AccountSettingsLayout() {
-  return (
-    <div className="flex w-full flex-col">
-      <div className="mx-auto w-full max-w-7xl pr-6 pb-16 pl-3 lg:px-8">
-        <div className="pt-2 pb-8">
-          <h1 className="pl-3 font-medium font-pp text-3xl text-foreground tracking-tight">
-            Your Account
-          </h1>
-        </div>
-
-        <div className="flex gap-12">
-          <SettingsSidebar
-            items={[
-              { name: "General", to: "/account/settings/general" },
-              {
-                name: "Source Control & Git",
-                to: "/account/settings/source-control",
-              },
-            ]}
-          />
-
-          <div className="min-w-0 max-w-4xl flex-1">
-            <Outlet />
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
